+++ archive-gpu/ch05/sphere.h
@@ -13,7 +13,7 @@ public:
 
 public:
   point3 center;
-  double radius;
+  float radius;
 };
 
 __device__ bool sphere::hit(const ray & r, float t_min, float t_max, hit_record & rec) const {
